🧪 test: move key_event handling test to integration tests
