+++ src/home/tests.rs
@@ -12,72 +12,9 @@ impl Home {
 }
 
 #[test]
-fn test_key_events() {
-    let home = Home::mock();
-
-    let key_events = [
-        (
-            KeyEvent::new(KeyCode::Char('c'), KeyModifiers::CONTROL),
-            Some(Action::Quit),
-            "Test default quit event.",
-        ),
-        (
-            KeyEvent::from(KeyCode::Tab),
-            Some(Action::SelectForward),
-            "Test default pane switch.",
-        ),
-        (
-            KeyEvent::new(KeyCode::Tab, KeyModifiers::ALT),
-            Some(Action::SelectBackward),
-            "Test pane switch backwards",
-        ),
-        (
-            KeyEvent::from(KeyCode::Char('i')),
-            Some(Action::EnterInsert),
-            "Test entering insert mode.",
-        ),
-        (
-            KeyEvent::from(KeyCode::Esc),
-            Some(Action::EnterNormal),
-            "Test exiting normal mode.",
-        ),
-        (
-            KeyEvent::from(KeyCode::Enter),
-            Some(Action::Submit),
-            "Test submitting in normal mode. (Should still retun the submit action)",
-        ),
-        (
-            KeyEvent::from(KeyCode::Char('j')),
-            Some(Action::NavigateDown),
-            "Test navigating table down with j.",
-        ),
-        (
-            KeyEvent::from(KeyCode::Down),
-            Some(Action::NavigateDown),
-            "Test navigating table down with DOWN.",
-        ),
-        (
-            KeyEvent::from(KeyCode::Char('k')),
-            Some(Action::NavigateUp),
-            "Test navigating table up with k.",
-        ),
-        (
-            KeyEvent::from(KeyCode::Up),
-            Some(Action::NavigateUp),
-            "Test navigating table up with UP.",
-        ),
-    ];
-
-    key_events.into_iter().for_each(|(event, want, desc)| {
-        let got = home.handle_key_events(event);
-        assert_eq!(want, got, "{desc}")
-    });
-}
-
-#[test]
 fn test_component_cycling_forward() {
     let mut test_home = Home::mock();
-    let key_event = KeyEvent::from(KeyCode::Tab);
+    let action = Some(Action::SelectForward);
 
     assert_eq!(
         test_home.component_tracker.inner(),
@@ -88,7 +25,6 @@ fn test_component_cycling_forward() {
     for i in 0..100 {
         let want = if i % 2 == 0 { 1 } else { 0 };
 
-        let action = test_home.handle_events(Some(Event::Key(key_event)));
         test_home.handle_action(action);
         assert_eq!(
             want,
@@ -101,7 +37,7 @@ fn test_component_cycling_forward() {
 #[test]
 fn test_component_cycling_backwards() {
     let mut test_home = Home::mock();
-    let key_event = KeyEvent::new(KeyCode::Tab, KeyModifiers::ALT);
+    let action = Some(Action::SelectBackward);
 
     assert_eq!(
         test_home.component_tracker.inner(),
@@ -112,7 +48,6 @@ fn test_component_cycling_backwards() {
     for i in 0..100 {
         let want = if i % 2 == 0 { 1 } else { 0 };
 
-        let action = test_home.handle_events(Some(Event::Key(key_event)));
         test_home.handle_action(action);
         assert_eq!(
             want,
@@ -124,16 +59,15 @@ fn test_component_cycling_backwards() {
 
 #[test]
 fn test_tracker_sync() {
-    let mut home = Home::mock();
-    let key_event = KeyEvent::from(KeyCode::Tab);
+    let mut test_home = Home::mock();
+    let action = Some(Action::SelectForward);
 
     for i in 0..100 {
-        let action = home.handle_events(Some(Event::Key(key_event)));
-        home.handle_action(action);
+        test_home.handle_action(action);
 
         let want = if i % 2 == 0 { false } else { true };
 
-        let [item, receipts] = &home.components[..] else {
+        let [item, receipts] = &test_home.components[..] else {
             panic!("Test case should only have two components.");
         };
 
@@ -143,7 +77,7 @@ fn test_tracker_sync() {
 }
 
 #[test]
-fn handle_response() -> Result<()> {
+fn handle_error_response() -> Result<()> {
     let res = DbResponse::new().req_type(RequestType::Update).error(
         RequestError::missing_param(RequestType::Update, "item", "id"),
     );
